requirements.txt
python-dotenv
Flask-PyMongo
pymongo
Flask
google-generativeai
flask-cors